backend/ml/deployment/requirement.txt
azureml-core==1.55.0
azureml-mlflow==1.55.0
mlflow==2.13.0
